Respond to some review
